Remove "capitalize initial term" feature.
This is required by the test suite but not the spec.

It makes more sense for us to do this capitalization in
the calling program, e.g. pandoc.  For some citations
in note styles may already be in notes and thus not
trigger separate footnotes.  If initial terms had
been capitalized, we'd need to uncapitalize, and that
is hard to do reliably.