Fix sorting on dates.
Previously this broke for some styles, e.g. apa.csl, which styles
dates as MM/DD/YYYY, and would lead to incorrect sorting of dates
with months and/or days.  With this change we sort on a normalized
date YYYY-MM-DD (though using only the components that would be
rendered), as described in the spec.

(As a trick to handle proper sorting of positive and negative dates,
we add 5000 internally when sorting the year.)

Closes #126.