CSL JSON: allow formatting in numeric fields.
See #170.

There's a catch, though. Currently the number splitting code
(`splitNums`) has to convert everything to text, so the
formatting will be lost. Still, this is better than treating
the formatting code as plain text which will then be escaped
in the output.  So, for example, we get
`1er` instead of `1&#60;sup&#62;er&#60;/sup&#62;` for
`1<sup>er</sup>`.