Fix handling of CSL JSON formatting.
By CSL JSON formatting, I mean tags like `<i>` and also quotes,
which are "smartified."

We were previously handling this in the wrong way, by parsing
it in the `fromText` method for `CslJson Text`.  This meant that
there was no way to prevent quotes from turning into smart
quotes, even in things like URLs (see #173).

What is more, we were not handling this sort of formatting at
all in other contexts, e.g. in terms, `text` elements with a
`value`, locators, prefixes, and suffixes.  The CSL tests were
passing only because the raw HTML was being passed through
verbatim, but in other formats this led to undesirable results
(see #170).

In this rather big patch, we revise how this is handled.
Formatting is now parsed in prefix, suffix, locator, text
value, and terms.  And `fromText` is now just `CslText`, so that
purely textual values like URLs can be inserted.

We also handle the conversion of Unicode superscript characters
to explicit `<sup>.</sup>` in CSL JSON somewhat differently.
Instead of making the replacement in *parsing* CSL JSON,
which would now affect other formats too, because terms in
locales are parsed as CSL JSON, we make the replacement in
*rendering* CSL JSON. This gives us behavior consistent with
the previous behavior, with the substitution occuring in
CSL JSON output but not in Pandoc.

[API changes]

+ Citeproc.Types now exports `parseCslJson`, `lookupQuotes`,
  `superscriptChars`.
+ Citeproc.Locale no longer exports `lookupQuotes`.
+ `parseCslJson` now takes an additional parameter for a Locale.

Closes #170.
Closes #173.

I think that this also gives us a path for #163 and #68 (which
seem to be the same); we could have pandoc render the
`citationLocator` string as CSL JSON (instead of using
`stringify`). The CSL JSON formatting would then be parsed by
citeproc.